+++ examples/src/bin/embassy_serial.rs
@@ -89,8 +89,10 @@ async fn main(spawner: Spawner) {
 
     let config = Config::default().with_rx_fifo_full_threshold(READ_BUF_SIZE as u16);
 
-    let mut uart0 = Uart::new(peripherals.UART0, config, rx_pin, tx_pin)
+    let mut uart0 = Uart::new(peripherals.UART0, config)
         .unwrap()
+        .with_tx(tx_pin)
+        .with_rx(rx_pin)
         .into_async();
     uart0.set_at_cmd(AtCmdConfig::default().with_cmd_char(AT_CMD));