+++ hil-test/tests/uart_async.rs
@@ -27,8 +27,10 @@ mod tests {
 
         let (rx, tx) = hil_test::common_test_pins!(peripherals);
 
-        let uart = Uart::new(peripherals.UART0, uart::Config::default(), rx, tx)
+        let uart = Uart::new(peripherals.UART0, uart::Config::default())
             .unwrap()
+            .with_tx(tx)
+            .with_rx(rx)
             .into_async();
 
         Context { uart }